nixOS handled in a better way as its own OS

+++ shell.nix
@@ -1,7 +1,8 @@
 # pinning nixpkgs in version 21.11 (https://github.com/NixOS/nixpkgs/releases/tag/21.11), done as explained here: https://nix.dev/tutorials/towards-reproducibility-pinning-nixpkgs
-{ pkgs ? import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/ce6aa13369b667ac2542593170993504932eb836.tar.gz") {
-  config.allowUnfree = true; # for steam-run
-} }:
+{ pkgs ? import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/06278c77b5d162e62df170fec307e83f1812d94b.tar.gz") {
+    # config.allowUnfree = true; # for steam-run
+  }
+}:
 
 let
   my-python = pkgs.python3;
@@ -19,7 +20,7 @@ pkgs.mkShell {
   buildInputs = [
     python-with-my-packages
     pkgs.cacert # to fix certificate issues (https://github.com/NixOS/nixpkgs/issues/66716#issuecomment-883399373)
-    pkgs.steam-run # I know this sucks, but making nix work sucks more..
+    # pkgs.steam-run # I know this sucks, but making nix work sucks more..
     #pkgs.curl
     #pkgs.unzip
     #pkgs.tar
@@ -31,8 +32,22 @@ pkgs.mkShell {
     # Download dependencies using the Python script
     python3 download_dependencies.py
 
-    mkdir ./bin/nixos/
-    cp ${pkgs.syncthing}/bin/syncthing ./bin/nixos/
-    cp ${(pkgs.callPackage ./nix/playit-cli.nix {})}/bin/playit-cli ./bin/nixos/
+    if [ ! -d ./bin/nixos/syncthing ]; then
+      echo "[shell.nix]: Downloading syncthing..."
+      mkdir -p ./bin/nixos/syncthing/
+      cp ${pkgs.syncthing}/bin/syncthing ./bin/nixos/syncthing/
+    else
+      echo "[shell.nix]: Directory ./bin/nixos/syncthing already exists, skipping download"
+    fi
+
+    if [ -f "./bin/nixos/playit-cli" ]; then
+      echo "[shell.nix]: File ./bin/nixos/playit-cli already exists, skipping download."
+    else
+      echo "[shell.nix]: Downloading playit-cli..."
+      cp ${(pkgs.callPackage ./nix/playit-cli.nix {})}/bin/playit-cli ./bin/nixos/
+    fi
+
+    echo
+    echo "run 'python3 main.py' to start the program"
   '';
 }
